Split playlist switcher and tabs preferences into two tabs
