+++ foo_ui_columns/config.h
@@ -27,7 +27,8 @@ preferences_tab* g_get_tab_artwork();
 preferences_tab* g_get_tab_display2();
 preferences_tab* g_get_tab_pview_artwork();
 preferences_tab* g_get_tab_sys();
-preferences_tab* g_get_tab_playlist();
+preferences_tab* g_get_tab_playlist_switcher();
+preferences_tab* g_get_tab_playlist_tabs();
 preferences_tab* g_get_tab_playlist_dd();
 preferences_tab* g_get_tab_main();
 preferences_tab* g_get_tab_status();
